restructure a bit more. reorg internal modules. filesystem paths for wallets are boken. Next commit to resolve this
