+++ unnamedwallet/utils/__init__.py
@@ -1,0 +1,4 @@
+from algodinstance import algodinstance 
+import constants
+
+name = "utils"
