--- unnamedwallet/walletcore/utils/__init__.py
@@ -1,3 +1,0 @@
-from . import algodinstance # This still isn't good reference. Find a better way @Todo
-
-name = "utils"
