+++ unnamedwallet/walletcore/wip_old/AssetBalance.py
