+++ unnamedwallet/walletcore/wip_old/ConsolidateWalletBalance.py
